PR fixes, refactoring, fix typos
