+++ driver-core/src/main/com/mongodb/internal/function/CheckedConsumer.java
@@ -14,19 +14,19 @@
  * limitations under the License.
  */
 
-package com.mongodb.internal;
+package com.mongodb.internal.function;
 
 /**
  * <p>This class is not part of the public API and may be removed or changed at any time</p>
  */
 @FunctionalInterface
-public interface CheckedSupplier<T, E extends Exception> {
+public interface CheckedConsumer<T, E extends Exception> {
 
     /**
-     * Gets a result.
+     * Performs this operation on the given argument.
      *
-     * @return a result
+     * @param t the input argument
      * @throws E the checked exception to throw
      */
-    T get() throws E;
+    void accept(T t) throws E;
 }
