+++ driver-core/src/main/com/mongodb/internal/function/CheckedFunction.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.mongodb.internal;
+package com.mongodb.internal.function;
 
 /**
  * <p>This class is not part of the public API and may be removed or changed at any time</p>
