+++ driver-core/src/main/com/mongodb/internal/function/CheckedRunnable.java
@@ -14,10 +14,18 @@
  * limitations under the License.
  */
 
+package com.mongodb.internal.function;
+
 /**
+ * <p>This class is not part of the public API and may be removed or changed at any time</p>
  */
+@FunctionalInterface
+public interface CheckedRunnable<E extends Exception> {
 
-@NonNullApi
-package com.mongodb.internal.async;
-
-import com.mongodb.lang.NonNullApi;
+    /**
+     * Checked run.
+     *
+     * @throws E the checked exception to throw
+     */
+    void run() throws E;
+}
