+++ driver-core/src/main/com/mongodb/internal/function/CheckedSupplier.java
@@ -14,18 +14,19 @@
  * limitations under the License.
  */
 
-package com.mongodb.internal;
+package com.mongodb.internal.function;
 
 /**
  * <p>This class is not part of the public API and may be removed or changed at any time</p>
  */
 @FunctionalInterface
-public interface CheckedRunnable<E extends Exception> {
+public interface CheckedSupplier<T, E extends Exception> {
 
     /**
-     * Checked run.
+     * Gets a result.
      *
+     * @return a result
      * @throws E the checked exception to throw
      */
-    void run() throws E;
+    T get() throws E;
 }
