+++ driver-core/src/main/com/mongodb/internal/package-info.java
@@ -15,7 +15,6 @@
  */
 
 /**
- * This package contains classes that manage binding to MongoDB servers for various operations.
  */
 
 @NonNullApi
