+++ driver-reactive-streams/src/test/unit/com/mongodb/reactivestreams/client/internal/gridfs/GridFSUploadPublisherImplTest.java
@@ -93,9 +93,7 @@ class GridFSUploadPublisherTest {
 
             Throwable throwable = onErrorEvents.get(0);
             assertEquals(MongoOperationTimeoutException.class, throwable.getClass());
-            assertEquals(
-                    // TODO-CSOT move all messages to fields?
-                    "GridFS timed out waiting for data from provided source Publisher", throwable.getMessage());
+            assertEquals("GridFS timed out waiting for data from provided source Publisher", throwable.getMessage());
 
             //assert no chunk has been inserted as we have not sent any data from source publisher.
             for (CommandEvent event : commandListener.getEvents()) {
