+++ driver-sync/src/main/com/mongodb/client/internal/CommandMarker.java
@@ -43,8 +43,7 @@ import static com.mongodb.internal.capi.MongoCryptHelper.startProcess;
 
 @SuppressWarnings("UseOfProcessBuilder")
 class CommandMarker implements Closeable {
-    // TODO-CSOT what is difference? Is a Command Marker part of an operation? (creates an extra method)
-    private static final String TIMEOUT_ERROR_MESSAGE = "Command marking exceeded the timeout limit.";
+    private static final String TIMEOUT_ERROR_MESSAGE = "Command marker exceeded the timeout limit.";
     @Nullable
     private final MongoClient client;
     @Nullable
